requirements.txt
filetype
httpx
asyncio
flask
flask-sqlalchemy
waitress
pyrogram
psycopg2
tgcrypto
tiktoken
nest_asyncio
langdetect
indic-nlp-library
